+++ app/graphql/queries/orders/orders_query.rb
@@ -3,7 +3,7 @@
 module Queries
   module Orders
     class OrdersQuery < BaseQuery
-      argument :input, Types::Custom::Inputs::Filtrations::OrdersFiltrationInput, required: false
+      argument :input, Types::Custom::Inputs::Filtrations::Orders::OrdersInput, required: false
       type Types::Custom::Objects::Orders::OrderWithAllQuantityObject, null: false
 
       def resolve(params)
