--- app/graphql/types/custom/inputs/filtrations/opinion_filtration_input.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  module Custom
-    module Inputs
-      module Filtrations
-        class OpinionFiltrationInput < Types::BaseInputObject
-          argument :pagination, PaginationInput, required: false
-        end
-      end
-    end
-  end
-end
