+++ app/graphql/types/custom/inputs/filtrations/opinions/opinion_input.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Types
+  module Custom
+    module Inputs
+      module Filtrations
+        module Opinions
+          class OpinionInput < Types::BaseInputObject
+            argument :pagination, PaginationInput, required: false
+          end
+        end
+      end
+    end
+  end
+end
