--- app/graphql/types/custom/inputs/filtrations/orders_filtration_input.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  module Custom
-    module Inputs
-      module Filtrations
-        class OrdersFiltrationInput < Types::BaseInputObject
-          argument :user_id, ID, required: false
-          argument :pagination, PaginationInput, required: true
-        end
-      end
-    end
-  end
-end
